pyproject.toml
[tool.poetry]
name = "nft-data-pull"
version = "0.1.0"
description = ""
authors = ["tomfutago <tomfutago@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
pandas = "^1.5.3"
web3 = "^6.0.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
